requirements.txt
requests
numpy>=2.3
pandas>=1.5
google-api-python-client
beautifulsoup4
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
plotly
pandas